Make OSX (homebrew) friendly
The OSX version of `nc` doesn't correctly support the -w timeout value, in
that failed connections take up to 1 minute to time out regradless of
the timer value.

Using the `ncat` version installed from the homebrew nmap package honors
the -w timeout paramter, so default to that and fallabck to `nc`

The pattern matching string used to strip `http://` or `https://` doesn't
work in the latest bash version, installed by brew. To optionally match
the "s" it needs to use the extglob pattern `?(s)`, which matches 0 or 1
"s" characters, the old `?` matches a single character thus `http://`
does not match. For this to work, extglog needs to be enabled in the
script.